requirements.txt
# Core Apify dependencies
apify >= 2.0.0

# Browser automation with nodriver
nodriver >= 0.37

# Data validation and parsing - compatible with Crawlee
pydantic >= 2.6.0

# Date and time utilities
python-dateutil >= 2.8.2

# HTTP client for API calls
aiohttp >= 3.9.0

# URL parsing and utilities
urllib3 >= 2.0.0

# Async utilities
asyncio-throttle >= 1.0.2

# Logging and debugging
coloredlogs >= 15.0.1

# Data processing
lxml >= 5.0.0
beautifulsoup4 >= 4.12.0

# Type hints support
typing-extensions >= 4.9.0